client/requirements.txt
aiohttp
pydantic
python-dotenv
pyyaml
